fix variable type

+++ repo/lib/my/my_strstr.c
@@ -19,7 +19,7 @@ char *my_strstr(char *str, char const *to_find)
 {
     size_t stop = my_strlen(str) - my_strlen(to_find);
 
-    for (int i = 0; i <= stop; i++)
+    for (size_t i = 0; i <= stop; i++)
         if (check(str + i, to_find))
             return str + i;
     return NULL;
